requirements.txt
Flask
findspark
pyspark
kafka-python
confluent-kafka[avro]